moving user class into models package
